Add files route. Save player data. Various updates. Code clean up.

+++ utils/storage.ts
@@ -30,8 +30,13 @@ export function setLocalStorage(key: string, value: unknown) {
   }
 }
 
-export function deleteLocalStorage() {
+export function deleteLocalStorage(key?: string) {
   if (import.meta.client) {
+    if (key) {
+      window.localStorage.removeItem(key);
+      return;
+    }
+
     window.localStorage.clear();
   }
 }
